refactor(users): change folder structure
