+++ src/app/users/[username]/style.module.scss
@@ -16,15 +16,3 @@
   margin-top: 40px;
   border-top: 1px solid $grey400;
 }
-
-.tabs {
-  padding: 20px 0;
-  display: flex;
-  justify-content: center;
-  font-weight: 500;
-}
-
-.tab {
-  display: flex;
-  gap: 10px;
-}
